feat: investment control panel overview
